+++ .pre-commit-config.yaml
@@ -1,0 +1,18 @@
+repos:
+- repo: local
+  hooks:
+    - id: system
+      name: black
+      entry: poetry run black
+      language: system
+      types: [python]
+    - id: system
+      name: isort
+      entry: poetry run isort
+      language: system
+      types: [python]
+    - id: system
+      name: flake9
+      entry: poetry run flake8
+      language: system
+      types: [python]
