+++ donguri_gaeru/core/utils.py
@@ -1,32 +1,26 @@
 # Copyright (C) 2021 mazziechai
-# 
+#
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU Affero General Public License as
 # published by the Free Software Foundation, either version 3 of the
 # License, or (at your option) any later version.
-# 
+#
 # This program is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
 # MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 # GNU Affero General Public License for more details.
-# 
+#
 # You should have received a copy of the GNU Affero General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-import logging
+import re
 
-from config import config
+bot = None  # This is to prevent recursive imports when something needs the bot instance
 
-logger = logging.getLogger("discord")
-logger.setLevel(config["log_level"])
 
-formatter = logging.Formatter("%(asctime)s:%(levelname)s:%(name)s: %(message)s")
-
-file_handler = logging.FileHandler(filename="dongurigaeru.log", encoding="utf-8", mode="w")
-file_handler.setFormatter(formatter)
-
-stream_handler = logging.StreamHandler()
-stream_handler.setFormatter(formatter)
-
-logger.addHandler(file_handler)
-logger.addHandler(stream_handler)
+def escape_markdown(text):
+    return re.sub(
+        r"\\\\([_*\[\]()~`>\#\+\-=|\.!])",
+        r"\1",
+        re.sub(r"([_*\[\]()~`>\#\+\-=|\.!])", r"\\\1", text),
+    )
